Fix log file testing to use directory and worker_id for log file
Instead of just using directory as the full path. This requires us to
create the folder using os.mkdir on the path that we get from twisted's
mktemp. Previously we would just use the path given by mktemp as the
full path for the file, instead of just the folder.